aider: Clear the sharing link
# Aider chat conversation:

USER: In this code i.e "Set-PnPListItemPermission -List $ListName -Identity $Item.Id -ClearSharingLink" the -ClearSharingLink does not exist.

ASSISTANT: I see, it seems like there's a misunderstanding about the `Set-PnPListItemPermission` cmdlet. The `-ClearSharingLink` parameter does not exist in this cmdlet.

If you want to clear the sharing link of a list item, you should use the `Set-PnPListItem` cmdlet with the `-Values` parameter to set the `SharedWithUsersId` field to an empty array.

Here's how you can modify the code:

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
                #Clear the sharing link expiration
                Set-PnPListItemPermission -List $ListName -Identity $Item.Id -ClearSharingLink
=======
                #Clear the sharing link
                Set-PnPListItem -List $ListName -Identity $Item.Id -Values @{"SharedWithUsersId" = @()}
>>>>>>> updated
```

Please note that this will only clear the sharing link if the item was shared with specific users. If the item was shared with a link (anyone with the link can access), you might need to use the `Remove-PnPSharingLink` cmdlet.