aider: Added optional parameter `$FolderName` to specify a folder within the document library or list.
# Aider chat conversation:

USER: Question: removingsharingfileaccess does that remove the ability for the file to be shared?

ASSISTANT: No, the `$RemoveSharingFileAccess` parameter does not remove the ability for the file to be shared. It only removes the existing shared access links to the file.

In other words, if `$RemoveSharingFileAccess` is set to `true`, the script will remove all existing shared links to the file, effectively making it so that anyone who had access to the file through those links will no longer be able to access it. However, new shared links can still be created for the file in the future.

If `$RemoveSharingFileAccess` is set to `false`, the script will not remove any existing shared links to the file.

USER: How does that affect the expiration date field for the file?

ASSISTANT: The `$RemoveSharingFileAccess` parameter and the expiration date of a shared link are two separate things.

When you create a shared link for a file in SharePoint, you have the option to set an expiration date for the link. This means the link will stop working after the specified date, and anyone trying to access the file through the link will no longer be able to do so.

If `$RemoveSharingFileAccess` is set to `true`, the script will remove all existing shared links to the file, regardless of whether they have an expiration date or not. This means that even if a link was set to expire in the future, it will be removed immediately when the script is run.

If `$RemoveSharingFileAccess` is set to `false`, the script will not remove any existing shared links to the file. This means that any links with an expiration date will continue to work until the date they are set to expire.

So, to summarize, the `$RemoveSharingFileAccess` parameter controls whether existing shared links to the file are removed immediately when the script is run. It does not directly affect the expiration date of the shared links. However, if a link is removed, it will no longer work, regardless of its expiration date.

USER: that sounds exactly what I need, I test this but is it possible to select a folder within the document library or listname; could this folder be an optional parameter

ASSISTANT: Yes, we can add an optional parameter `$FolderName` to specify a folder within the document library or list. If this parameter is provided, the script will only process the items within this folder. If it's not provided, the script will process all items in the document library or list.

Here are the changes we need to make:

1. Add a new parameter `$FolderName` to the parameters list.
2. Modify the line where we get the list items to check if `$FolderName` is provided and, if so, only get the items within this folder.

Here are the proposed changes:

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
#Parameters
param(
    [Parameter(Mandatory=$true)]
    [string]$SiteUrl,
    [Parameter(Mandatory=$true)]
    [string]$ReportOutput,
    [Parameter(Mandatory=$true)]
    [string]$ListName,
    [Parameter(Mandatory=$false)]
    [string]$LogFile = "log.txt",
    [Parameter(Mandatory=$false)]
    [bool]$RemoveSharingFileAccess = $true
)
=======
#Parameters
param(
    [Parameter(Mandatory=$true)]
    [string]$SiteUrl,
    [Parameter(Mandatory=$true)]
    [string]$ReportOutput,
    [Parameter(Mandatory=$true)]
    [string]$ListName,
    [Parameter(Mandatory=$false)]
    [string]$LogFile = "log.txt",
    [Parameter(Mandatory=$false)]
    [bool]$RemoveSharingFileAccess = $true,
    [Parameter(Mandatory=$false)]
    [string]$FolderName
)
>>>>>>> updated
```

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
#Get all list items in batches
#Only fetch the fields we need to improve performance
$ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
=======
#Get all list items in batches
#Only fetch the fields we need to improve performance
if ($null -ne $FolderName) {
    $ListItems = Get-PnPListItem -List $ListName -Folder $FolderName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
} else {
    $ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
}
>>>>>>> updated
```

These changes will add a new parameter `$FolderName` and will only process the items within the specified folder if this parameter is provided.