aider: aider: Added optional parameter $FolderName to specify a folder.
# Aider chat conversation:

USER: Can you provide an update on the `README.md` based on the changes done so far

ASSISTANT: Sure, I can help with that. However, I need to know what changes have been made to the project so far. Could you please provide me with the details of the changes you've made?

USER: This is based on the commit that was done recently

ASSISTANT: I'm sorry for the confusion, but as an AI, I don't have access to your commit history. Could you please provide the details of the changes that were made in the recent commit? This will help me to update the `README.md` file accurately.

USER: https://github.com/vkanimea/get-sp-dl-report/commits/main

ASSISTANT: I'm sorry for any misunderstanding, but as an AI, I don't have the ability to access external websites or repositories, including GitHub. However, I can guide you on how to update the `README.md` file based on the changes you describe. Please provide the details of the changes made in the recent commit, and I'll help you draft the update.

USER: aider: Added optional parameter $FolderName to specify a folder wit…

ASSISTANT: Based on your description, it seems like an optional parameter `$FolderName` has been added to specify a folder. Here's how you can update the `README.md` file to reflect this change:

```markdown
README.md
<<<<<<< HEAD
# get-sp-dl-report
# Generate a report for a Sharepoint Site for expired sharelinks in a Document library
# Parameters are SiteURL, ReportOutput, ListName {Document Library}
# Uses PnP Modules for Sharepoint
# ReportOutput: csv file output : Name, RelativeURL, FileType, ShareLink, ShareLinkAccess, ShareLinkType, AllowsAnonymousAccess, IsActive, Expiration
# To Do
# 1. External Parameters for SiteURL, ReportOutput, ListName
# 2. Add External Parameter to remove Expiration
=======
# get-sp-dl-report
# Generate a report for a Sharepoint Site for expired sharelinks in a Document library
# Parameters are SiteURL, ReportOutput, ListName {Document Library}, FolderName {Optional - Specify a folder}
# Uses PnP Modules for Sharepoint
# ReportOutput: csv file output : Name, RelativeURL, FileType, ShareLink, ShareLinkAccess, ShareLinkType, AllowsAnonymousAccess, IsActive, Expiration
# To Do
# 1. External Parameters for SiteURL, ReportOutput, ListName, FolderName
# 2. Add External Parameter to remove Expiration
>>>>>>> updated
```

This edit block adds the `$FolderName` parameter to the list of parameters and the list of external parameters to be added in the future.