aider: Fixed an issue where the `FolderServerRelativeUrl` argument was not being checked for null or empty values before passing it to the `Get-PnPListItem` cmdlet.
# Aider chat conversation:

USER: An unknown error occurred: The 'query.FolderServerRelativeUrl' argument is invalid.

ASSISTANT: The error message suggests that the `FolderServerRelativeUrl` argument passed to the `Get-PnPListItem` cmdlet is invalid. This could be due to a few reasons:

1. The `$FolderName` variable is null or empty.
2. The `$FolderName` variable does not contain a valid server-relative URL.

To fix this, we need to add a check to ensure that `$FolderName` is not null or empty before passing it to the `Get-PnPListItem` cmdlet. Here's how you can modify your script:

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
    #Get all list items in batches
    #Only fetch the fields we need to improve performance
    if ($null -ne $FolderName) {
        $ListItems = Get-PnPListItem -List $ListName -Folder $FolderName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
    } else {
        $ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
    }
=======
    #Get all list items in batches
    #Only fetch the fields we need to improve performance
    if (![string]::IsNullOrEmpty($FolderName)) {
        $ListItems = Get-PnPListItem -List $ListName -Folder $FolderName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
    } else {
        $ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
    }
>>>>>>> updated
```

This change will ensure that the `Get-PnPListItem` cmdlet is only called with the `-Folder` parameter if `$FolderName` is not null or empty.