aider: The `$ReportOutput` parameter is now optional with a default value of `report.csv`.
# Aider chat conversation:

USER: Can the $ReportOutput parameter not be mandatory and output a file similar to the log.txt

ASSISTANT: Yes, we can make the `$ReportOutput` parameter optional and set a default value similar to the `$LogFile` parameter. If no value is provided, the script will output the report to a file named `report.csv` in the same directory as the script.

Here's the proposed change:

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
    [Parameter(Mandatory=$true)]
    [string]$ReportOutput,
=======
    [Parameter(Mandatory=$false)]
    [string]$ReportOutput = "report.csv",
>>>>>>> updated
```

This edit block will change the `$ReportOutput` parameter from mandatory to optional and set a default value of `report.csv`.