aider: Updated the example command in the README.md file to reflect the new default value of the `$RemoveSharingFileAccess` parameter.
# Aider chat conversation:

USER: Can we update the README.md based on the changes so far.

ASSISTANT: Sure, we can update the README.md to reflect the changes made to the script. Here's the proposed change:

<source>markdown
README.md
<<<<<<< HEAD
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ReportOutput "C:\path\to\your\report.csv" -ListName "Your List Name" -RemoveSharingFileAccess $true
=======
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ReportOutput "C:\path\to\your\report.csv" -ListName "Your List Name" -RemoveSharingFileAccess $false
>>>>>>> updated
</source>

This edit block will change the example command in the README.md file to reflect the new default value of the `$RemoveSharingFileAccess` parameter.