aider: Added support for using a previously generated output file to improve processing speed.
# Aider chat conversation:

USER: Is it possible to use a previous output file that was generated earllier to run against. This means that the processing will be faster as use the existing output instead of running against the documents in the Document library.

ASSISTANT: Yes, it is possible to use a previously generated output file to run the script against. This would indeed make the processing faster as the script would not need to fetch the documents from the Document Library again.

To implement this, we would need to add a new parameter to the script, say `$UsePreviousOutput`, which would be a boolean. If `$UsePreviousOutput` is set to `$true`, the script would read the data from the previously generated output file (specified by `$ReportOutput`) instead of fetching the documents from the Document Library.

Here's the proposed change:

<source>powershell
get-sp-rp.ps1
<<<<<<< HEAD
    [Parameter(Mandatory=$false)]
    [bool]$RemoveSharingFileAccess = $false
=======
    [Parameter(Mandatory=$false)]
    [bool]$RemoveSharingFileAccess = $false,
    [Parameter(Mandatory=$false)]
    [bool]$UsePreviousOutput = $false
>>>>>>> updated
</source>

This edit block will add the `$UsePreviousOutput` parameter to the script.

Next, we need to modify the part of the script that fetches the documents from the Document Library to only do so if `$UsePreviousOutput` is `$false`. If `$UsePreviousOutput` is `$true`, the script should read the data from the output file.

<source>powershell
get-sp-rp.ps1
<<<<<<< HEAD
    #Get all list items in batches
    #Only fetch the fields we need to improve performance
    $ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
    $ItemCount = $ListItems.Count
=======
    #Get all list items in batches or from previous output
    if ($UsePreviousOutput -eq $false) {
        #Only fetch the fields we need to improve performance
        $ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
        $ItemCount = $ListItems.Count
    } else {
        $ListItems = Import-Csv -Path $ReportOutput
        $ItemCount = $ListItems.Count
    }
>>>>>>> updated
</source>

This edit block will modify the script to fetch the documents from the Document Library only if `$UsePreviousOutput` is `$false`. If `$UsePreviousOutput` is `$true`, the script will read the data from the output file.