aider: Updated the README.md file to mention that the `-ReportOutput` and `-LogFile` parameters log to a default file and only need to be set if the user wants to change the default directory for the output files.
# Aider chat conversation:

USER: For the usage, the -ReportOutput and -Log

ASSISTANT: I'm sorry, but your message seems to be incomplete. Could you please provide more details or clarify your request? Are you asking for changes in the usage instructions for the `-ReportOutput` and `-LogFile` parameters in the `README.md` file?

USER: For the usage, the non mandatory parameters such as -ReportOutput and -LogFile already output to a default file so don't need to add this in the usage just report that it is log to default file. Only set it if you need to change the default directory for the output files

ASSISTANT: Understood. We need to modify the `README.md` file to update the usage instructions for the `-ReportOutput` and `-LogFile` parameters. We will remove the specific examples for these parameters and instead mention that they log to a default file, and only need to be set if the user wants to change the default directory for the output files.

Here is the proposed change:

<source>markdown
README.md
<<<<<<< HEAD
If you want to specify a log file, you can do so with the `-LogFile` parameter:
```powershell
.\get-sp-rp.ps1 -SiteUrl "https://yoursharepointsite.com" -ReportOutput "C:\path\to\your\report.csv" -ListName "Your List Name" -DestinationLibrary "Your Destination Library" -LogFile "C:\path\to\your\log.txt"
```
=======
By default, the script logs to a default file. If you want to change the default directory for the output files, you can do so with the `-LogFile` and `-ReportOutput` parameters.
>>>>>>> updated
</source>